app/Backend/requirements.txt
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distlib==0.3.8
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
google==3.0.0
google-api-core==2.20.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==3.0.0
platformdirs==4.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
# pyldap==3.0.0.post1
# python-ldap==3.4.4
# python3-ldap==0.9.8.4
python-dotenv
google-cloud-aiplatform
SQLAlchemy
validator-collection
PyMySQL
PyJWT
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
rsa==4.9
simple-websocket==1.0.0
soupsieve==2.6
urllib3==2.2.3
virtualenv
Werkzeug==3.0.4
wsproto==1.2.0
